fix: move envs to common
